add disable pagination to workflowinstance get
